Create guess.cpp
This is a simple number guessing game implemented in C++. The program generates a random number between 1 and 10 and prompts the user to guess the number. The user has a limited number of attempts to guess the correct number. The program provides feedback to the user by indicating whether their guess is too high or too low.